raspi/requirements.txt
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.3
pyserial==3.5
pyyaml==6.0.1
loguru==0.7.2

